# Conflicts:
#	AnimationStub/AnimationStub/Bone_PS.csh
#	AnimationStub/AnimationStub/Bone_VS.csh
#	AnimationStub/AnimationStub/FBXLoader.cpp
#	AnimationStub/AnimationStub/FBX_PS.csh
#	AnimationStub/AnimationStub/FBX_VS.csh
#	AnimationStub/AnimationStub/OBJ_PS.csh
#	AnimationStub/AnimationStub/OBJ_VS.csh
#	AnimationStub/AnimationStub/boxBind.cpp

+++ AnimationStub/AnimationStub/KeyFrame.h
@@ -25,7 +25,7 @@ public:
 private:
 	// When in the animation does this KeyFrame happen?
 	FbxTime keyTime;
-	float keyFrameNum;
+	int keyFrameNum;
 	FbxAMatrix globalTransform;
 	KeyFrame* next;
